api/cgen/requirements.txt
google-genai
python-dotenv
pydantic
fastapi
uvicorn
typing-extensions